requirements.txt
absl-py==0.4.1
apt-clone==0.2.1
apturl==0.5.2
astor==0.7.1
astroid==1.6.5
autopep8==1.4
beautifulsoup4==4.6.0
Brlapi==0.6.6
certifi==2018.8.24
chardet==3.0.4
command-not-found==0.3
configobj==5.0.6
cupshelpers==1.0
cycler==0.10.0
defer==1.0.6
future==0.16.0
gast==0.2.0
grpcio==1.14.2
gym==0.10.5
h5py==2.8.0
httplib2==0.9.2
idna==2.7
isort==4.3.4
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.3
mccabe==0.6.1
numpy==1.15.1
onboard==1.4.1
PAM==0.4.2
pexpect==4.2.1
Pillow==5.1.0
protobuf==3.6.1
psutil==5.4.2
pycairo==1.16.2
pycodestyle==2.4.0
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.1
pyglet==1.3.2
pygobject==3.26.1
PyICU==1.9.8
pyinotify==0.9.6
pylint==1.9.3
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.2.1
pyRFC3339==1.0
python-apt==1.6.0
python-dateutil==2.7.3
python-debian==0.1.32
python-xapp==1.2.0
python-xlib==0.20
pytz==2018.5
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.19.1
requests-unixsocket==0.1.5
selenium==3.14.1
sessioninstaller==0.0.0
setproctitle==1.1.10
six==1.11.0
system-service==0.3
systemd-python==234
tensorboard==1.10.0
tensorflow==1.10.1
termcolor==1.1.0
ubuntu-drivers-common==0.0.0
ufw==0.35
urllib3==1.23
Werkzeug==0.14.1
wrapt==1.10.11
xkit==0.0.0
